requirements.txt
asn1crypto==0.24.0
autoflake==1.2
autopep8==1.4
certifi==2018.8.24
cffi==1.11.5
cryptography==2.3.1
docformatter==1.0
emoji==0.5.1
ephem==3.7.6.0
future==0.16.0
idna==2.7
pycodestyle==2.4.0
pycparser==2.18
pyflakes==2.0.0
PySocks==1.6.8
python-telegram-bot==11.1.0
six==1.11.0
unify==0.4
untokenize==0.1.1
